Consolidate all test infrastructure into single test.js file
- Combine relay, chat client, and test runner into unified test.js
- Delete redundant files (chat.js, test-relay.js, test-runner.js, run-tests.sh)
- Support all test modes via command line arguments:
  - `npm test` - runs all tests (js-js, swift-swift, js-swift)
  - `npm run test:js-js` - test JS-to-JS only
  - `npm run test:swift-swift` - test Swift-to-Swift only
  - `npm run test:js-swift` - test JS-to-Swift only
  - `npm run relay` - run relay only
  - `npm run chat` - interactive chat mode
- Simplify CI workflow to just run `npm test` for all tests
- Built-in Nostr relay implementation
- Automatic cleanup and process management

This provides a single, self-contained test file that handles all
testing scenarios with proper npm script integration.
